me when i refactor frfr because i wanna do smth but i cant figure out the popup ui thing ah ha ha
