+++ src/app/trade/page.tsx
@@ -2,21 +2,13 @@
 
 import React from 'react';
 import {
-  Avatar,
-  Card, Flex,
+  Flex,
   HStack,
-  VStack,
-  Stack,
-  Text,
   Box,
   Heading,
   Slider,
 } from '@chakra-ui/react';
-import TradingCards from '@/components/trading/TradingCards';
-import {
-  LuStar, LuArrowRightLeft
-} from 'react-icons/lu';
-
+import TradeCard from '@/components/trading/TradeCard';
 
 type TradeCardProps = {
   username: string;
@@ -24,109 +16,69 @@ type TradeCardProps = {
   rating: number;
 };
 
-const TradeCard: React.FC<TradeCardProps> = ({ username, avatarUrl, rating}) =>{
-  return (
-    <Card.Root width="80%">
-      <Card.Body>
-          {/*their cards*/}
-          <VStack align="center" gap={2}>
-            <TradingCards />
-            <VStack align="center" gap="0">
-              <HStack mb="0" gap="3">
-                <Avatar.Root>
-                  <Avatar.Image src={avatarUrl ?? "https://images.unsplash.com/photo-1511806754518-53bada35f930"} />
-                  <Avatar.Fallback name={username} />
-                </Avatar.Root>
-                <Stack gap="0">
-                  <Text fontWeight="semibold" textStyle="sm">
-                    {username}
-                  </Text>
-                </Stack>
-              </HStack>
-              <Stack gap="0">
-                {/* show star and numeric rating side-by-side */}
-                <HStack gap="1" align="center">
-                  <Box>
-                    {(() => {
-                      const color =
-                        rating <= 2.5
-                          ? '#ff3b30'
-                          : rating < 4.0
-                            ? '#ffd60a'
-                            : rating < 5
-                              ? '#32d74b'
-                              : '#08a9c6';
-                      return <LuStar color={color} size={20} />;
-                    })()}
-                  </Box>
-                  <Text fontSize="sm" fontWeight="semibold">{Number.isFinite(rating) ? rating.toFixed(1) : '-'}</Text>
-                </HStack>
-              </Stack>
-            </VStack>
-          </VStack>
-
-
-          {/*/!*TODO: GET RID OF THIS, JUST SHOW THEIR CARDS, HAVE A POPUP WHEN YOU PRESS ON THE CARD*/}
-          {/*SHOWING WHAT THE TRADE IS, WHEN CLICK ON THEIR PFP OR USER LEAD TO THEIR PROFILE*!/*/}
-          {/*<LuArrowRightLeft size={150} />*/}
-
-          {/*/!*your cards*!/*/}
-          {/*<VStack align="center" gap={2}>*/}
-          {/*  <TradingCards />*/}
-          {/*  <Text fontSize="sm" fontWeight="semibold">Your cards</Text>*/}
-          {/*</VStack>*/}
-
-      </Card.Body>
-    </Card.Root>
-  )
-};
-
 const TradePage = () => {
-    const users: TradeCardProps[] = [
-        { username: 'Nate Foss', rating: 4.5 },
-        { username: 'Ava Johnson', rating: 4.0 },
-        { username: 'Liam Smith', rating: 1 },
-        { username: 'Maya Patel', rating: 3.5 },
-        { username: 'Carlos Ruiz', rating: 2 },
-        { username: 'Zoe Kim', rating: 5 },
-    ];
+  const users: TradeCardProps[] = [
+    { username: 'Nate Foss', rating: 4.5 },
+    { username: 'Ava Johnson', rating: 4.0 },
+    { username: 'Liam Smith', rating: 1 },
+    { username: 'Maya Patel', rating: 3.5 },
+    { username: 'Carlos Ruiz', rating: 2 },
+    { username: 'Zoe Kim', rating: 5 },
+  ];
 
-    return (
-        <Box bg="white" minH="100vh" color="black" mb={4}>
-            <Flex flexDirection="column" alignItems="center" gap={2}>
-                <Flex
-                    flexDirection="row"
-                    justifyContent="center"
-                    alignItems="center"
-                    gap={1}
-                >
-                </Flex>
-                <Box w="100%" position="relative" px={4}>
-                    <Heading mt={1} fontSize="2xl" textAlign="center" fontWeight={'Bold'} fontFamily="var(--font-sans)" maxW="container.md" mx="auto">
-                        TradePost
-                    </Heading>
+  return (
+    <Box bg="white" minH="100vh" color="black" mb={4}>
+      <Flex flexDirection="column" alignItems="center" gap={2}>
+        <Flex
+          flexDirection="row"
+          justifyContent="center"
+          alignItems="center"
+          gap={1}
+        ></Flex>
+        <Box w="100%" position="relative" px={4}>
+          <Heading
+            mt={1}
+            fontSize="2xl"
+            textAlign="center"
+            fontWeight={'Bold'}
+            fontFamily="var(--font-sans)"
+            maxW="container.md"
+            mx="auto"
+          >
+            TradePost
+          </Heading>
 
-                    <Box position="absolute" right={4} top="50%" transform="translateY(-50%)">
-                        <Slider.Root maxW="sm" size="sm" defaultValue={[40]}>
-                            <HStack justify="space-between">
-                                <Slider.Label>Distance</Slider.Label>
-                                <Slider.ValueText />
-                            </HStack>
-                            <Slider.Control>
-                                <Slider.Track>
-                                    <Slider.Range />
-                                </Slider.Track>
-                                <Slider.Thumbs />
-                            </Slider.Control>
-                        </Slider.Root>
-                    </Box>
-                </Box>
-                {users.map(u => (
-                    <TradeCard key={u.username} username={u.username} avatarUrl={u.avatarUrl} rating={u.rating} />
-                ))}
-            </Flex>
+          <Box
+            position="absolute"
+            right={4}
+            top="50%"
+            transform="translateY(-50%)"
+          >
+            <Slider.Root maxW="sm" size="sm" defaultValue={[40]}>
+              <HStack justify="space-between">
+                <Slider.Label>Distance</Slider.Label>
+                <Slider.ValueText />
+              </HStack>
+              <Slider.Control>
+                <Slider.Track>
+                  <Slider.Range />
+                </Slider.Track>
+                <Slider.Thumbs />
+              </Slider.Control>
+            </Slider.Root>
+          </Box>
         </Box>
-    )
+        {users.map((u) => (
+          <TradeCard
+            key={u.username}
+            username={u.username}
+            avatarUrl={u.avatarUrl}
+            rating={u.rating}
+          />
+        ))}
+      </Flex>
+    </Box>
+  );
 };
 
 export default TradePage;
